requirements.txt
Flask>=2.0
Flask-SQLAlchemy
Flask-Login
Flask-WTF
python-dotenv
Werkzeug
